Debugging, improved accuracy computation
Edited so that instructions appear in the figure window. Accuracy is actually being recorded now (I had to edit the logic so that waitforkeypress cmd was used for the correct_quadrant function. Still cannot figure out what is going on with reaction time??? Still matches duration. Definitiely something wrong with my logic but unsure what.