Final product?
I got rid of the instructions so that my matrices will display immediately after yours when we run them in succession. I also got rid of the duration like we talked about in class. The accuracy and reaction times seem to be working now. I also set it so that the target stimulus disappears as soon as a key is pressed.